requirement.txt
dashscope==1.15.0
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-openai==0.2.3
openai==1.52.0

pandas==1.5.3
sqlalchemy==2.0.29
PyMySQL==1.1.1
PyYAML==6.0.1

torch~=2.2.1
transformers~=4.39.1

pywebio
